feature [modif-api]: fix error handling for uploads
