+++ server/routes/routesWithDb.js
@@ -1,41 +1,14 @@
 /* eslint-disable no-console */
 import express from 'express';
-import multer from 'multer';
-import cloudinary from 'cloudinary';
-import cloudinaryStorage from 'multer-storage-cloudinary';
 import userHandler from '../controllers/dbUserHandler';
 import dbIncidentHandler from '../controllers/dbIncidentHandler';
 import incidentInputValidator from '../middlewares/incidentInputValidator';
 import userInputValidator from '../middlewares/userInputValidator';
 import authMiddleware from '../middlewares/authMiddleware';
 import mediaMiddleware from '../middlewares/mediaMiddleware';
-import dotEnv from '../config/config';
 
 const app = express.Router();
 const apiVersion = '/api/v2/';
-
-cloudinary.config({
-  cloud_name: process.env.CLOUD_NAME || dotEnv.CLOUD_NAME,
-  api_key: process.env.API_KEY || dotEnv.API_KEY,
-  api_secret: process.env.API_SECRET || dotEnv.API_SECRET,
-});
-
-const storage = cloudinaryStorage({
-  cloudinary,
-  folder: 'iReporter/media',
-  allowedFormats: ['jpg', 'svg', 'png', 'jpeg', 'gif', 'avi', 'flv', 'mpeg', '3gp', 'mp4'],
-  // allowedFormats: (req, file, cb) => {
-  //   if (!file.mimetype.includes('image') && !file.mimetype.includes('vidoe')) {
-  //     return cb(new Error(`${file.mimetype} is not supported`));
-  //   }
-  //   return '';
-  // },
-});
-
-const upload = multer({
-  storage,
-  limits: 100000,
-});
 
 app
   .route(`${apiVersion}user/create`)
@@ -57,7 +30,7 @@ app
   .route(`${apiVersion}incident/create`)
   .post(
     authMiddleware.validateToken,
-    upload.any(),
+    mediaMiddleware.multerCheck,
     incidentInputValidator.createIncidentQueryValidator,
     mediaMiddleware.returnedFiles,
     dbIncidentHandler.createIncident,
